Fix round bug

+++ my_coffee.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users(
   id INTEGER PRIMARY KEY AUTOINCREMENT,
   name TEXT,
-  credits DOUBLE DEFAULT 0,
-  credits_spent DOUBLE DEFAULT 0
+  credits INTEGER DEFAULT 0,
+  credits_spent INTEGER DEFAULT 0
 );
